see http://code.google.com/p/wiquery/issues/detail?id=95
svn path=/branches/1.0.x/; revision=422

+++ src/test/java/org/odlabs/wiquery/ui/core/options/OptionsTestCase.java
@@ -78,7 +78,7 @@ public class OptionsTestCase extends TestCase{
 		options.put("keyInt", 1);
 		
 		Assert.assertFalse(options.isEmpty());
-		Assert.assertEquals(1, options.getInt("keyInt"));
+		Assert.assertEquals(1, options.getInt("keyInt").intValue());
 	}
 	
 	/*@Test
